# perf(cubesql): Make incremental rule scheduler distinguish between generation 0 and 1